Moved main() into separate module
