+++ isd_s3/__init__.py
@@ -7,6 +7,8 @@ __all__ = (
     "isd_s3"
 )
 
+__version__ = "0.1.4"
+
 """
     Configure logging for a library, per Python best practices:
     https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library
